requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
WTForms>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
email-validator>=2.0.0
Pillow>=9.0.0
python-dateutil>=2.8.0
Flask-Mail>=0.9.0
itsdangerous>=2.0.0 
python-docx>=0.8.11
PyPDF2>=3.0.0